Re: #495 - BUG: Videos don't play correctly in acceptance
remote_video_mime_type returns a value no matter the mime type so
i changed the logic and allowed anything to pass except text/html.
this is still fragile as it will allow any other mime type thru
and rely on mediaelementplayer to handle it, which it may not be able
to do, however, it’s leniency allows for other video mime types.

another approach would be to create a method that only allows
specific mime types and pass a hash containing the mime type detected
and whether or not it should be diplayed by mediaelementplayer