Re: #516 - BUG: microsoft translator
re-factored the front-end code so that it no longer requests a new token.
the method is unreliable since the ms response cannot be parsed because
the ms api responds with an error message that creates a js paser error.
the main area where this causes a problem is when the original text is
in a language not handled by the ms translator, e.g. serbian.

also removed languages available look-up via the api in order to save on
http overhead. rely now on the values cached in microsoft.translator